feat(editor2): complete operator/group controls
